[phi] Split selected_rows CMake compilation
